feat(middlewared): move EventSource to its own module
Ticket:	#26278
